В конструктор и деструктор класса поместил функции инициализации `EVP` из `main`.
