+++ src/crypto_guard_ctx.cpp
@@ -1,6 +1,7 @@
 #include "crypto_guard_ctx.h"
 
 #include <array>
+#include <vector>
 #include <openssl/evp.h>
 
 namespace CryptoGuard {
@@ -17,8 +18,46 @@ struct AesCipherParams {
 
 class CryptoGuardCtx::PImpl {
 public:
-    PImpl() = default;
-    ~PImpl() = default;
+    PImpl() {
+        std::string input = "01234567890123456789";
+
+        OpenSSL_add_all_algorithms();
+
+        auto params = CreateChiperParamsFromPassword("12341234");
+        params.encrypt = 1;
+        ctx_ = EVP_CIPHER_CTX_new();
+
+        outBuf_.resize(16 + EVP_MAX_BLOCK_LENGTH);
+
+        // Инициализируем cipher
+        EVP_CipherInit_ex(ctx_, params.cipher, nullptr, params.key.data(), params.iv.data(), params.encrypt);
+
+        std::vector<unsigned char> inBuf(16);
+
+        // Обрабатываем первые N символов
+        std::copy(input.begin(), std::next(input.begin(), 16), inBuf.begin());
+        EVP_CipherUpdate(ctx_, outBuf_.data(), &outLen_, inBuf.data(), static_cast<int>(16));
+        for (int i = 0; i < outLen_; ++i) {
+            output_.push_back(outBuf_[i]);
+        }
+
+        // Обрабатываем оставшиеся символы
+        std::copy(std::next(input.begin(), 16), input.end(), inBuf.begin());
+        EVP_CipherUpdate(ctx_, outBuf_.data(), &outLen_, inBuf.data(), static_cast<int>(input.size() - 16));
+        for (int i = 0; i < outLen_; ++i) {
+            output_.push_back(outBuf_[i]);
+        }
+    }
+    ~PImpl() {
+        // Заканчиваем работу с cipher
+        EVP_CipherFinal_ex(ctx_, outBuf_.data(), &outLen_);
+        for (int i = 0; i < outLen_; ++i) {
+            output_.push_back(outBuf_[i]);
+        }
+        EVP_CIPHER_CTX_free(ctx_);
+        std::print("String encoded successfully. Result: '{}'\n\n", output_);
+        EVP_cleanup();
+    }
 
     PImpl(const PImpl &) = delete;
     PImpl &operator=(const PImpl &) = delete;
@@ -30,6 +69,27 @@ public:
     void DecryptFile(std::iostream &inStream, std::iostream &outStream, std::string_view password) {}
 
     std::string CalculateChecksum(std::iostream &inStream) { return "NOT_IMPLEMENTED"; }
+
+private:
+    AesCipherParams CreateChiperParamsFromPassword(std::string_view password) {
+        AesCipherParams params;
+        constexpr std::array<unsigned char, 8> salt = {'1', '2', '3', '4', '5', '6', '7', '8'};
+
+        int result = EVP_BytesToKey(params.cipher, EVP_sha256(), salt.data(),
+                                    reinterpret_cast<const unsigned char *>(password.data()), password.size(), 1,
+                                    params.key.data(), params.iv.data());
+
+        if (result == 0) {
+            throw std::runtime_error{"Failed to create a key from password"};
+        }
+
+        return params;
+    }
+
+    EVP_CIPHER_CTX *ctx_;
+    std::vector<unsigned char> outBuf_;
+    int outLen_;
+    std::string output_;
 };
 
 CryptoGuardCtx::CryptoGuardCtx() : pImpl_(nullptr) {}
